+++ pkg/ffuf/util.go
@@ -3,6 +3,9 @@ package ffuf
 import (
 	"math/rand"
 	"os"
+	"os/exec"
+	"strconv"
+	"strings"
 )
 
 //used for random string generation in calibration function
@@ -40,4 +43,31 @@ func FileExists(path string) bool {
 	}
 
 	return !md.IsDir()
+}
+
+//SizeofTTY return two integers or an error (X, Y, error) that corresspond to the width and height of the TTY and any error that could occur.
+func SizeofTTY() (int, int, error) {
+	cmd := exec.Command("stty", "size")
+	cmd.Stdin = os.Stdin
+	out, err := cmd.Output()
+	if err != nil {
+		return 0, 0, err
+	}
+	output := string(out)
+
+	parts := strings.Split(output, " ")
+	x, err := strconv.Atoi(parts[0])
+	if err != nil {
+		return 0, 0, err
+	}
+	y, err := strconv.Atoi(strings.Replace(parts[1], "\n", "", 1))
+	if err != nil {
+		return 0, 0, err
+	}
+	return x, y, nil
+}
+
+//Percent take current completed jobs and total jobs and return the percentage completed
+func Percent(current int, total int) float32 {
+	return ((float32(current) * float32(100)) / float32(total))
 }